+++ examples/github/policies/policy.yaml
@@ -1,87 +1,81 @@
 ---
 # sample policy for validating repositories
 version: v1
-type: pipeline-policy
+type: policy
 name: acme-github-policy
 context:
   group: Root Group
   provider: github
 repository:
-  - context: github
-    rules:
-      - type: secret_scanning
-        def:
-          enabled: true
-      - type: secret_push_protection
-        def:
-          enabled: true
-      - type: branch_protection
-        params:
-          branch: main
-        def:
-          required_pull_request_reviews:
-            dismiss_stale_reviews: true
-            require_code_owner_reviews: true
-            required_approving_review_count: 1
-          required_linear_history: true
-          allow_force_pushes: false
-          allow_deletions: false
-          allow_fork_syncing: true
-      - type: github_actions_allowed
-        def:
-          allowed_actions: selected
-      - type: allowed_selected_actions
-        def:
-          github_owned_allowed: true
-          verified_allowed: true
-          patterns_allowed: []
-      - type: default_workflow_permissions
-        def:
-          default_workflow_permissions: read
-          can_approve_pull_request_reviews: false
-      - type: codeql_enabled
-        def: {}
-      - type: actions_check_pinned_tags
-        def: {}
-      - type: dependabot_configured
-        def:
-          package_ecosystem: gomod
-          schedule_interval: weekly
-          apply_if_file: go.mod
-      - type: dependabot_configured
-        def:
-          package_ecosystem: npm
-          schedule_interval: weekly
-          apply_if_file: package.json
-      - type: dockerfile_no_latest_tag
-        def: {}
-      - type: trivy_action_enabled
-        def: {}
+  - type: secret_scanning
+    def:
+      enabled: true
+  - type: secret_push_protection
+    def:
+      enabled: true
+  - type: branch_protection
+    params:
+      branch: main
+    def:
+      required_pull_request_reviews:
+        dismiss_stale_reviews: true
+        require_code_owner_reviews: true
+        required_approving_review_count: 1
+      required_linear_history: true
+      allow_force_pushes: false
+      allow_deletions: false
+      allow_fork_syncing: true
+  - type: github_actions_allowed
+    def:
+      allowed_actions: selected
+  - type: allowed_selected_actions
+    def:
+      github_owned_allowed: true
+      verified_allowed: true
+      patterns_allowed: []
+  - type: default_workflow_permissions
+    def:
+      default_workflow_permissions: read
+      can_approve_pull_request_reviews: false
+  - type: codeql_enabled
+    def: {}
+  - type: actions_check_pinned_tags
+    def: {}
+  - type: dependabot_configured
+    def:
+      package_ecosystem: gomod
+      schedule_interval: weekly
+      apply_if_file: go.mod
+  - type: dependabot_configured
+    def:
+      package_ecosystem: npm
+      schedule_interval: weekly
+      apply_if_file: package.json
+  - type: dockerfile_no_latest_tag
+    def: {}
+  - type: trivy_action_enabled
+    def: {}
 artifact:
-  - context: github
-    rules:
-      - type: artifact_signature
-        params:
-          tags: [main]
-          name: test
-        def:
-          is_signed: true
-          is_verified: true
-          is_bundle_verified: true
+  - type: artifact_signature
+    params:
+      tags: [main]
+      name: test
+    def:
+      is_signed: true
+      is_verified: true
+      is_bundle_verified: true
 pull_request:
-  - context: github
-    rules:
-      - type: pr_vulnerability_check
-        def:
-          action: reject_pr
-          ecosystem_config:
-            - name: npm
-              vulnerability_database_type: osv
-              vulnerability_database_endpoint: https://api.osv.dev/v1/query
-              package_repository:
-                url: https://registry.npmjs.org
-            - name: go
-              vulnerability_database_type: osv
-              vulnerability_database_endpoint: https://vuln.go.dev
-              package_repository:
-                url: https://proxy.golang.org
+  - type: pr_vulnerability_check
+    def:
+      action: reject_pr
+      ecosystem_config:
+        - name: npm
+          vulnerability_database_type: osv
+          vulnerability_database_endpoint: https://api.osv.dev/v1/query
+          package_repository:
+            url: https://registry.npmjs.org
+        - name: go
+          vulnerability_database_type: osv
+          vulnerability_database_endpoint: https://vuln.go.dev
+          package_repository:
+            url: https://proxy.golang.org
