+++ examples/github/policies/policy_artifact.yaml
@@ -1,19 +1,17 @@
 ---
 # sample policy for validating artifact signatures
 version: v1
-type: pipeline-policy
+type: policy
 name: acme-github-policy-artifact
 context:
   group: Root Group
   provider: github
 artifact:
-  - context: github
-    rules:
-      - type: artifact_signature
-        params:
-          tags: [main]
-          name: test
-        def:
-          is_signed: true
-          is_verified: true
-          is_bundle_verified: true
+  - type: artifact_signature
+    params:
+      tags: [main]
+      name: test
+    def:
+      is_signed: true
+      is_verified: true
+      is_bundle_verified: true
